+++ config/simulation_config.json
@@ -1,0 +1,62 @@
+{
+    "FPS": 60,
+    "SCALE": 5.0,
+    "SIM_WIDTH": 1300,
+    "SIM_HEIGHT": 800,
+    "SIM_SIZE": [
+        1300,
+        800
+    ],
+    "BACKGROUND_COLOR": [
+        0,
+        0,
+        0
+    ],
+    "GRID_COLOR": [
+        100,
+        100,
+        100
+    ],
+    "VEHICLE_COLOR": [
+        255,
+        255,
+        255
+    ],
+    "TIRE_COLOR": [
+        255,
+        255,
+        255
+    ],
+    "MARK_COLOR": [
+        80,
+        80,
+        80
+    ],
+    "HUD_BG_COLOR": [
+        50,
+        50,
+        50,
+        180
+    ],
+    "HUD_FG_COLOR": [
+        255,
+        255,
+        0
+    ],
+    "GRAVITY": 9.81,
+    "AIR_DENSITY": 1.225,
+    "DRAG_COEFF": 0.3,
+    "ROLL_RESIST": 0.015,
+    "PHYSICS_SUBSTEPS": 4,
+    "ENABLE_TRACK_MARKS": true,
+    "ENABLE_DEBUG_INFO": true,
+    "CAMERA_FOLLOW": true,
+    "TERRAIN_FRICTION": {
+        "asphalt": 1.0,
+        "gravel": 0.8,
+        "grass": 0.7,
+        "dirt": 0.6,
+        "snow": 0.3,
+        "ice": 0.1
+    }
+}
